amazon-scrapper/pyproject.toml
[tool.poetry]
name = "amazon-scrapper"
version = "0.1.0"
description = ""
authors = ["Marcel Stolin <mstolin@mailbox.org>"]

[tool.poetry.scripts]
start = "amazon_scrapper.main:start"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.74.1"
uvicorn = {extras = ["standard"], version = "^0.17.5"}
beautifulsoup4 = "^4.10.0"
lxml = "^4.8.0"
requests = "^2.27.1"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
